requirements.txt
# Core ML dependencies
torch>=1.9.0
tensorflow>=2.11.0
tensorflow_probability>=0.19.0
fairseq>=0.12.0
transformers>=4.33.2
accelerate>=0.23.0
diffusers>=0.21.2
timm>=0.6.13
tensorboard>=2.6.0
tensorboardX>=2.6
wandb>=0.12.0  # Optional: for experiment tracking

# Data processing and utilities
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.61.0
scikit-learn>=0.24.2
scikit-image>=0.16.2
h5py>=3.1.0
PyYAML>=5.4.1
typeguard>=2.13.3
einops>=0.6.1
hydra-core>=1.0.7

# MIDI and Music Processing
midiutil>=1.2.1
miditoolkit>=1.0.1
pretty-midi>=0.2.10
pypianoroll>=1.0.0
basic-pitch>=0.2.5
ddsp>=1.6.0
demucs>=4.0.0
librosa>=0.8.0

# Audio Processing
pyloudnorm>=0.1.1
webrtcvad>=2.0.10
praat-parselmouth>=0.3.3

# Text Processing
datasets>=2.11.0
textgrid>=1.5
jieba>=0.42.1
pypinyin>=0.39.0
g2pM>=0.1.2.5
cnsenti>=0.0.7
textblob>=0.17.1
jiwer>=3.0.1
inflect>=5.3.0

# UI/Visualization
gradio>=3.26.0
pytesseract>=0.3.10